spike/requirements.txt
openreward
openai
pydantic>=2
python-dotenv
